Crud-pessoa

+++ src/main/java/com/emagalha/desafio_api/dto/PageDTO.java
@@ -6,12 +6,12 @@ import java.util.List;
 @Data
 public class PageDTO<T> {
     private List<T> content;
-    private long totalElements;
+    private Integer totalElements;
     private int totalPages; 
     private int currentPage;  
     private int pageSize; 
 
-    public PageDTO(List<T> content, long totalElements, int totalPages, int currentPage, int pageSize) {
+    public PageDTO(List<T> content, Integer totalElements, int totalPages, int currentPage, int pageSize) {
         this.content = content;
         this.totalElements = totalElements;
         this.totalPages = totalPages;
